requirements.txt
# Reddit API (primary scraping method)
praw>=7.7.0

# HTTP requests (for API and utilities)
requests>=2.31.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
sqlite3  # Built-in with Python

# Database connectors
mysql-connector-python>=8.2.0
PyMySQL>=1.1.0
SQLAlchemy>=2.0.0

# NLP and text processing
spacy>=3.7.0
nltk>=3.8.0
textblob>=0.17.0
langdetect>=1.0.9
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0

# Machine learning and deep learning
scikit-learn>=1.3.0
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0

# Text analysis and keyword extraction
keybert>=0.8.0
yake>=0.4.8
textstat>=0.7.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
colorama>=0.4.6
pyyaml>=6.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Optional: For advanced features
# openai>=1.0.0  # For GPT integration
# anthropic>=0.7.0  # For Claude integration
